moved IextensionExporter file toi shared layer
