+++ 3ds Max/Max2Babylon/2017/Max2Babylon2017.csproj
@@ -164,9 +164,6 @@
     <Compile Include="..\Exporter\BabylonExporter.cs">
       <Link>Exporter\BabylonExporter.cs</Link>
     </Compile>
-    <Compile Include="..\Exporter\BabylonExporter.IExtensionExporter.cs">
-      <Link>Exporter\BabylonExporter.IExtensionExporter.cs</Link>
-    </Compile>
     <Compile Include="..\Exporter\BabylonExporter.IMaterialExporter.cs">
       <Link>Exporter\BabylonExporter.IMaterialExporter.cs</Link>
     </Compile>
