+++ 3ds Max/Max2Babylon/2018/Max2Babylon2018.csproj
@@ -163,9 +163,6 @@
     <Compile Include="..\Exporter\BabylonExporter.cs">
       <Link>Exporter\BabylonExporter.cs</Link>
     </Compile>
-    <Compile Include="..\Exporter\BabylonExporter.IExtensionExporter.cs">
-      <Link>Exporter\BabylonExporter.IExtensionExporter.cs</Link>
-    </Compile>
     <Compile Include="..\Exporter\BabylonExporter.IMaterialExporter.cs">
       <Link>Exporter\BabylonExporter.IMaterialExporter.cs</Link>
     </Compile>
