+++ SharedProjects/BabylonExport.Entities/BabylonScene.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.IO;
 using System.Runtime.Serialization;
+using Babylon2GLTF;
 using GLTFExport.Entities;
 using Max2Babylon;
